chore:Rename useSWR to useApi
