+++ src/lib/Api/useApi.svelte.ts
@@ -1,10 +1,10 @@
-import type { SWRKey } from './types'
+import type { ApiPathKey } from './types'
 import { defaultJsonFetcher } from './utils'
 
 export const useApi = <R>(path: string | undefined) => {
   let data: R | undefined = $state(undefined)
 
-  const fetcher = defaultJsonFetcher as (url: SWRKey) => Promise<R>
+  const fetcher = defaultJsonFetcher as (url: ApiPathKey) => Promise<R>
 
   if (path) {
     fetcher(path).then((result) => {
