+++ src/lib/Api/utils.ts
@@ -1,6 +1,6 @@
-import type { SWRKey } from './types'
+import type { ApiPathKey } from './types'
 
-export const defaultJsonFetcher = async <D>(url: SWRKey): Promise<D> => {
+export const defaultJsonFetcher = async <D>(url: ApiPathKey): Promise<D> => {
   const response = await fetch(url)
   if (!response.ok) throw new Error('Not a 2XX response.')
   return response.json()
